One type per file is better
