+++ Plugins/GitUIPluginInterfaces/GitUIPluginInterfaces.csproj
@@ -87,6 +87,8 @@
     <Compile Include="IGitRemoteCommand.cs" />
     <Compile Include="ILockableNotifier.cs" />
     <Compile Include="IProcess.cs" />
+    <Compile Include="SettingControlBinding.cs" />
+    <Compile Include="ISettingControlBinding.cs" />
     <Compile Include="ISettingsValueGetter.cs" />
     <Compile Include="ObjectId.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
